+++ src/cpp/tests/test_tardigrade_balance_of_energy.cpp
@@ -4,9 +4,9 @@
  * Tests for tardigrade_balance_equations_balance_of_energy
  */
 
+#include <tardigrade_LinearHex.h>
 #include <tardigrade_balance_of_energy.h>
 #include <tardigrade_constitutive_tools.h>
-#include <tardigrade_finite_element_utilities.h>
 #include <tardigrade_hydraLinearTestMaterial.h>
 #define USE_EIGEN
 #include <tardigrade_vector_tools.h>
@@ -2416,10 +2416,10 @@ void evaluate_at_nodes(const xi_in &xi_begin, const xi_in &xi_end, dt_type dt, c
                                    std::begin(u_dot_tp1), std::end(u_dot_tp1), dUDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p;
@@ -2567,10 +2567,10 @@ void evaluate_at_nodes(
                                    std::begin(u_dot_tp1), std::end(u_dot_tp1), dUDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p;
@@ -5550,10 +5550,10 @@ void evaluate_at_nodes(const xi_in &xi_begin, const xi_in &xi_end, dt_type dt, c
                                    std::begin(e_dot_tp1), std::end(e_dot_tp1), dEDotdE);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p,
@@ -5791,10 +5791,10 @@ void evaluate_at_nodes(
                                    std::begin(e_dot_tp1), std::end(e_dot_tp1), dEDotdE);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p,