+++ src/cpp/tests/test_tardigrade_balance_of_linear_momentum.cpp
@@ -4,9 +4,9 @@
  * Tests for tardigrade_balance_equations_balance_of_linear_momentum
  */
 
+#include <tardigrade_LinearHex.h>
 #include <tardigrade_balance_of_linear_momentum.h>
 #include <tardigrade_constitutive_tools.h>
-#include <tardigrade_finite_element_utilities.h>
 #include <tardigrade_hydraLinearTestMaterial.h>
 #define USE_EIGEN
 #include <tardigrade_vector_tools.h>
@@ -973,10 +973,10 @@ void evaluate_at_nodes(const xi_in &xi_begin, const xi_in &xi_end, dt_type dt, c
                                  std::end(u_ddot_tp1), dUDDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p;
@@ -1111,10 +1111,10 @@ void evaluate_at_nodes(const xi_in &xi_begin, const xi_in &xi_end, dt_type dt, c
                                  std::end(u_ddot_tp1), dUDDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p;
@@ -2612,10 +2612,10 @@ void evaluate_at_nodes(
                                  alpha, beta, std::begin(a_tp1), std::end(a_tp1), dUDDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p,
@@ -2849,10 +2849,10 @@ void evaluate_at_nodes(
                                  alpha, beta, std::begin(a_tp1), std::end(a_tp1), dUDDotdU);
 
     // Instantiate the element
-    tardigradeBalanceEquations::finiteElementUtilities::LinearHex<
+    tardigradeBalanceEquations::finiteElement::LinearHex<
         floatType, typename std::array<floatType, 24>::const_iterator,
         typename std::array<floatType, 3>::const_iterator, typename std::array<floatType, 8>::iterator,
-        typename std::array<floatType, 24>::iterator>
+        typename std::array<floatType, 24>::iterator, typename std::array<floatType, 3>::iterator, floatType>
         e(std::cbegin(x_tp1), std::cend(x_tp1), X_begin, X_end);
 
     std::array<typename std::iterator_traits<density_tp1_in>::value_type, nphases> density_tp1_p, density_dot_tp1_p,